[voicecall] Force critical telepathy accounts back online immediately on failure
Mission control implements reconnecting logic for accounts, but it has a
fairly severe backoff, because it's tuned for internet-connected
accounts. For telephony, it's very important that we get the account
back up as quickly as possible on failure to avoid missing incoming
traffic.

We can force mission control to reconnect the account instantly by
requesting an available presence immediately after failures. For
example, this causes telepathy-ring to be restarted instantly after
crashing or being killed.